PAE-1411: extract prn-summary factory and tighten shared types
